move cards components from seat and table to Card component, move shared func and variables to util and variables folder
