+++ app/components/Card/BoardCards.js
@@ -4,7 +4,7 @@ import Card from 'components/Card';
 import { BoardWrapper, CardWrapper } from './styles';
 
 // eslint-disable-next-line react/prefer-stateless-function
-class Board extends React.Component {
+class BoardCards extends React.Component {
   static propTypes = {
     board: PropTypes.array.isRequired,
   };
@@ -23,4 +23,4 @@ class Board extends React.Component {
   }
 }
 
-export default Board;
+export default BoardCards;
