+++ app/components/Table/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Board from 'components/Board';
+import BoardCards from 'components/Card/BoardCards';
 import TableMenu from '../../containers/TableMenu';
 import ActionBar from '../../containers/ActionBar';
 import tableImage from './tableBG.svg';
@@ -30,7 +30,7 @@ const TableComponent = (props) => (
 
           { props.seats }
 
-          <Board board={props.board} />
+          <BoardCards board={props.board} />
 
           { props.winners.length > 0 &&
             <Winner className="winner">{ props.winners }</Winner>
