+++ app/variables.js
@@ -82,3 +82,7 @@ export const curtainWidth = '400px';
 export const curtainHalfWidth = '200px';
 // Max width sticking curtain
 export const curtainStickyWidth = '1168px';
+
+// shadows
+export const medShadow = '0 2px 4px 0 rgba(0,0,0,0.50)';
+export const smallShadow = '0 1px 4px 0 rgba(0,0,0,0.50)';
